Big type refactor: moving around files and splitting into subfiles
