+++ optd-dsl/src/analyzer/mod.rs
@@ -3,6 +3,5 @@ pub mod errors;
 pub mod from_ast;
 pub mod hir;
 pub mod map;
-pub mod registry_check;
-pub mod type_infer;
+pub mod semantic_checks;
 pub mod types;
