--- optd-dsl/src/analyzer/type_infer/mod.rs
@@ -1,9 +1,0 @@
-mod generate;
-#[allow(unused)]
-mod glb;
-#[allow(unused)]
-mod lub;
-#[allow(unused)]
-mod solver;
-
-pub use solver::Solver;
